+++ packages/core-js-compat/src/data.mjs
@@ -2486,6 +2486,10 @@ export const data = {
   'esnext.iterator.to-array': null,
   'esnext.iterator.to-async': {
   },
+  'esnext.iterator.zip': {
+  },
+  'esnext.iterator.zip-keyed': {
+  },
   'esnext.json.is-raw-json': {
     bun: '1.1.43',
     chrome: '114',