+++ packages/core-js-compat/src/modules-by-versions.mjs
@@ -293,5 +293,7 @@ export default {
     'es.suppressed-error.constructor',
     'es.symbol.async-dispose',
     'es.symbol.dispose',
+    'esnext.iterator.zip',
+    'esnext.iterator.zip-keyed',
   ],
 };