+++ tests/eslint/eslint.config.js
@@ -1249,6 +1249,8 @@ const forbidCompletelyNonExistentBuiltIns = {
   'es/no-nonstandard-iterator-properties': [ERROR, { allow: [
     'concat',
     'range',
+    'zip',
+    'zipKeyed',
   ] }],
   'es/no-nonstandard-iterator-prototype-properties': [ERROR, { allow: [
     'toAsync',